gpu::position::lattice: fix conversion warning with GCC 4.1.
halmd/mdsim/gpu/position/lattice.cpp:127: warning: converting to ‘unsigned int’ from ‘double’

+++ halmd/mdsim/gpu/position/lattice.cpp
@@ -124,7 +124,9 @@ void lattice<dimension, float_type, RandomNumberGenerator>::set()
     }
     LOG("placing particles on fcc lattice: a = " << a);
 
-    unsigned int N = u * accumulate(n.begin(), n.end(), 1, multiplies<unsigned int>());
+    unsigned int N = static_cast<unsigned int>(
+        u * accumulate(n.begin(), n.end(), 1, multiplies<unsigned int>())
+    );
     if (N > particle->nbox) {
         LOG_WARNING("lattice not fully occupied (" << N << " sites)");
     }
